+++ package.json
@@ -6,7 +6,7 @@
     "scripts": {
         "test": "echo \"Error: no test specified\" && exit 1",
         "build": "webpack",
-        "serve": "webpack serve",
+        "serve": "webpack serve --open",
         "dev": "webpack-dev-server"
     },
     "keywords": [],
@@ -17,17 +17,21 @@
         "@babel/plugin-proposal-class-properties": "^7.18.6",
         "@babel/preset-env": "^7.18.9",
         "clean-webpack-plugin": "^4.0.0",
+        "copy-webpack-plugin": "^11.0.0",
         "css-loader": "^6.7.1",
+        "html-loader": "^4.1.0",
         "html-webpack-plugin": "^5.5.0",
+        "mini-css-extract-plugin": "^2.6.1",
         "postcss": "^8.4.14",
         "postcss-loader": "^7.0.1",
         "style-loader": "^3.3.1",
+        "svg-inline-loader": "^0.8.2",
         "webpack": "^5.73.0",
         "webpack-cli": "^4.10.0",
         "webpack-dev-server": "^4.9.3"
     },
     "dependencies": {
-        "autoprefixer": "^10.4.7",
-        "http-proxy-middleware": "^2.0.6"
+        "autoprefixer": "^10.4.7"
+
     }
 }
