+++ src/style.css
@@ -1,136 +1,1 @@
-* {
-    box-sizing: border-box;
-}
-
-.wrapper {
-    /*test*/
-    width: 600px;
-    /* max-height: 200px;  */
-    margin: 0 auto;
-    position: relative;
-}
-
-.slider {
-    background-color: rgb(192, 192, 192);
-    display: flex;
-    justify-content: space-evenly;
-    align-items: center;
-    max-height: 400px;
-}
-
-.slider__arrow-l {
-    border-top: 0.22rem solid black;
-    border-right: 0.22rem solid black;
-    width: 0.8rem;
-    height: 0.8rem;
-    transform: rotate(225deg);
-    margin: 0.15rem 0 0.15rem 0.3rem;
-}
-
-.slider__items {
-    width: 80%;
-    overflow: hidden;
-}
-
-.slider__visible {
-    /* gap: 2vw; */
-    justify-content: space-between;
-    display: flex;
-    flex-direction: row;
-    transition: cubic-bezier(0.55, 0.085, 0.68, 0.53) 0.5s;
-}
-
-.slider__item {
-    padding: 0.7rem;
-    background-color: azure;
-    border-radius: 30%;
-    display: flex;
-}
-
-.slider__item img {
-    height: 100%;
-}
-
-.slider__arrow-r {
-    border-top: 0.22rem solid black;
-    border-right: 0.22rem solid black;
-    width: 0.8rem;
-    height: 0.8rem;
-    transform: rotate(45deg);
-    margin: 0.15rem 0.3rem 0.15rem 0;
-}
-
-.hidden {
-    border-color: gray;
-    color: gray;
-}
-
-.arrow__r,
-.arrow__l {
-    padding: 0.7rem;
-    background-color: azure;
-    border-radius: 50%;
-    display: flex;
-    text-align: center;
-}
-
-.inputs {}
-
-input {
-    display: block;
-    margin-left: auto;
-    margin-right: auto;
-    margin-top: 20px;
-}
-
-.butt {
-    display: flex;
-    justify-content: center;
-    cursor: pointer;
-}
-
-.button {
-    margin-top: 20px;
-    padding: 10px;
-    border: 2px solid gray;
-    border-radius: 10px;
-    background-color: lightblue;
-    width: 150px;
-    text-align: center;
-}
-
-.hidden_s {
-    display: none;
-}
-
-body {
-    display: flex;
-    flex-direction: column;
-}
-
-.demo {
-    position: absolute;
-    width: 100%;
-    height: 100%;
-    background-color: RGBA(10, 29, 64, 0.6);
-    color: red;
-    text-transform: uppercase;
-    text-align: center;
-}
-
-.demo p {
-    background-color: black;
-    display: inline-block;
-    padding: 5px;
-    border-radius: 15px;
-}
-
-.demo p:nth-child(3) {
-    font-size: 60px;
-    margin: 0;
-}
-
-.demo__width {
-    min-width: 200px;
-    height: 200px;
-}
+@import "./slider/slider.css"
