--- src/template.html
@@ -1,62 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="style.css">
-    <title>Slider</title>
-</head>
-
-<body>
-    <div class="wrapper">
-        <div class="demo">
-            <p>Enter Value below, this is demo!</p><br>
-            <p>&#9759;</p>
-        </div>
-        <div class="slider ">
-            <div class="arrow__l">
-                <div class="slider__arrow-l"></div>
-            </div>
-
-            <div class="slider__items">
-                <div class="slider__visible">
-                    <div class="slider__item"><img src="../img/icons/1.svg" alt=""></div>
-                    <div class="slider__item"><img src="../img/icons/2.svg" alt=""></div>
-                    <div class="slider__item"><img src="../img/icons/3.svg" alt=""></div>
-                    <div class="slider__item"><img src="../img/icons/4.svg" alt=""></div>
-                    <div class="slider__item"><img src="../img/icons/5.svg" alt=""></div>
-                    <div class="slider__item"><img src="../img/icons/6.svg" alt=""></div>
-                    <div class="slider__item"><img src="../img/icons/7.svg" alt=""></div>
-                    <div class="slider__item"><img src="../img/icons/8.svg" alt=""></div>
-                    <div class="slider__item"><img src="../img/icons/9.svg" alt=""></div>
-                    <div class="slider__item"><img src="../img/icons/10.svg" alt=""></div>
-                </div>
-            </div>
-            <div class="arrow__r">
-                <div class="slider__arrow-r"></div>
-            </div>
-        </div>
-    </div>
-    <div class="inputs">
-        <input type="text" id="slides_to_show" placeholder="slides_to_show">
-        <input type="text" name="" id="slides_to_scroll" placeholder="slides_to_scroll">
-    </div>
-
-
-    <div class="butt" id="butt1">
-        <div class="button">
-            Apply
-        </div>
-    </div>
-
-    <div class="butt" id="butt2">
-        <div class="button">
-            Reset
-        </div>
-    </div>
-    <script src="./index.js"></script>
-</body>
-
-</html>
