+++ template.html
@@ -1,0 +1,68 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <!-- <link rel="stylesheet" href="style.css"> -->
+    <title>Slider</title>
+</head>
+
+<body>
+    <div class="wrapper">
+        <div class="demo">
+            <p>Enter Value below, this is demo!</p><br>
+            <p>&#9759;</p>
+        </div>
+        <div class="slider ">
+            <div class="arrow__l">
+                <div class="slider__arrow-l"></div>
+            </div>
+
+            <div class="slider__items">
+                <div class="slider__visible">
+                    <div class="slider__item"><img src="./src/assets/slider/1.svg" alt=""></div>
+                    <div class="slider__item"><img src="./src/assets/slider/2.svg" alt=""></div>
+                    <div class="slider__item"><img src="./src/assets/slider/3.svg" alt=""></div>
+                    <div class="slider__item"><img src="./src/assets/slider/4.svg" alt=""></div>
+                    <div class="slider__item"><img src="./src/assets/slider/5.svg" alt=""></div>
+                    <div class="slider__item"><img src="./src/assets/slider/6.svg" alt=""></div>
+                    <div class="slider__item"><img src="./src/assets/slider/7.svg" alt=""></div>
+                    <div class="slider__item"><img src="./src/assets/slider/8.svg" alt=""></div>
+                    <div class="slider__item"><img src="./src/assets/slider/9.svg" alt=""></div>
+                    <div class="slider__item"><img src="./src/assets/slider/10.svg" alt=""></div>
+                </div>
+            </div>
+            <div class="arrow__r">
+                <div class="slider__arrow-r"></div>
+            </div>
+        </div>
+    </div>
+    <!-- <div class="inputs">
+        <input type="text" id="slides_to_show" placeholder="slides_to_show">
+        <input type="text" name="" id="slides_to_scroll" placeholder="slides_to_scroll">
+    </div> -->
+
+
+
+
+    <div class="butt" id="butt1">
+        <div class="button">
+            Нажми Меня!
+        </div>
+    </div>
+
+    <form action="/">
+        <input type="text" id="slides_to_show" placeholder="slides_to_show">
+        <input type="text" name="" id="slides_to_scroll" placeholder="slides_to_scroll">
+        <div class="butt" id="butt1">
+            <input class="button" type="submit" value="Сброс">
+            </input>
+        </div>
+    </form>
+
+    <!-- <script src="./index.js"></script> -->
+</body>
+
+</html>
